GVR Unity SDK v1.40.0

+++ GoogleVR/Editor/GvrCompatibilityChecker.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2016 Google Inc. All rights reserved.
+// Copyright 2016 Google Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -31,6 +31,7 @@ using System.Linq;
 public class GvrCompatibilityChecker {
   // Asset subpaths.
   private static string PLUGINS_ANDROID_PATH = Application.dataPath + "/Plugins/Android/";
+  private static string PLUGINS_IOS_PATH = Application.dataPath + "/Plugins/iOS/";
   private static string ARMEABI_PATH = "libs/armeabi-v7a/";
   private static string X86_PATH = "libs/x86/";
   private static string IGNORE_MANIFEST_MERGE_CHECK_PATH = "GvrIgnoreManifestMergeCheck.txt";
@@ -39,13 +40,36 @@ public class GvrCompatibilityChecker {
   // Files for backwards compatibility.
   private static string ANDROID_MANIFEST = "AndroidManifest.xml";
   private static string ANDROID_MANIFEST_CARDBOARD = "AndroidManifest-Cardboard.xml";
-  private static string COMMON_AAR = "gvr_android_common.aar";
-  private static string GVR_ACTIVITY_AAR = "unitygvractivity.aar";
   private static string NATIVE_LIB = "libgvrunity.so";
+
+  private static string[] BACK_COMPAT_FILE_PATHS = new string[] {
+    PLUGINS_ANDROID_PATH + ARMEABI_PATH + NATIVE_LIB,
+    PLUGINS_ANDROID_PATH + X86_PATH + NATIVE_LIB,
+    PLUGINS_ANDROID_PATH + "gvr_android_common.aar",
+    PLUGINS_ANDROID_PATH + "unitygvractivity.aar",
+    PLUGINS_IOS_PATH + "libgvrunity.a",
+    PLUGINS_IOS_PATH + "CardboardAppController.h",
+    PLUGINS_IOS_PATH + "CardboardAppController.mm",
+  };
+
+  private static string[] BACK_COMPAT_DIR_PATHS =  new string[] {
+    PLUGINS_IOS_PATH + "CardboardSDK.bundle",
+    PLUGINS_IOS_PATH + "GoogleKitCore.bundle",
+    PLUGINS_IOS_PATH + "GoogleKitDialogs.bundle",
+    PLUGINS_IOS_PATH + "GoogleKitHUD.bundle",
+    PLUGINS_IOS_PATH + "MaterialRobotoFontLoader.bundle",
+  };
+
+  // Files for native integration compatibility.
+  private static string IOS_AUDIO_LIB = "libaudioplugingvrunity.a";
 
   // GVR backwards-compatible package.
   private static string BACK_COMPAT_PACKAGE_PATH =
     "/GoogleVR/GVRBackwardsCompatibility.unitypackage";
+
+  // iOS native integration-compatible package.
+  private static string IOS_NATIVE_COMPAT_PACKAGE_PATH =
+    "/GoogleVR/GVRiOSNativeCompatibility.unitypackage";
 
   // Path elements.
   private static string ASSET_PATH_PREFIX = "Assets";
@@ -58,8 +82,9 @@ public class GvrCompatibilityChecker {
     "The following files must be removed to avoid project compilation discrepancies:\n\n";
   private static string IMPORT_REQUIRED_TITLE = "Package Import Required";
   private static string IMPORT_REQUIRED_MESSAGE =
-    "Additional libraries must be imported for GVR to be compatible with " +
-    "this version of Unity, which does not have the GVR native integration.";
+    "Assets/GoogleVR/GVRBackwardsCompatibility.unitypackage must be imported for GVR to be " +
+    "compatible with this version of Unity, which does not have the GVR native integration.\n" +
+    "Please download this file from github.com/gvr-unity-sdk if it is not already in your project.";
   private static string MANIFEST_UPDATE_WARNING_TITLE = "AndroidManifest.xml Merge Required";
   private static string MERGE_MANIFEST_WARNING_MESSAGE =
     "Please merge the existing AndroidManifest.xml with AndroidManifest-Cardboard.xml.";
@@ -87,17 +112,33 @@ public class GvrCompatibilityChecker {
 #if !UNITY_HAS_GOOGLEVR
     if (!IgnoreCompatibilityCheck() &&
         !AllBackwardsCompatibilityFilesExist()) {
+      RemoveiOSNativeIntegrationFiles();
       ImportBackwardsCompatibilityPackage();
     }
 #else
     RemoveAnyBackwardsCompatibleFiles();
     AndroidManifestCompatibilityUpdate();
+#if UNITY_IOS
+    ImportiOSNativeCompatibilityPackage();
+#endif  // UNITY_IOS
 #endif  // !UNITY_HAS_GOOGLEVR
   }
 #endif  // UNITY_ANDROID || UNITY_IOS
 
   private static bool AllBackwardsCompatibilityFilesExist() {
-    return !GetBackCompatFilePaths().Where(filePath => !File.Exists(filePath)).Any();
+    return !BACK_COMPAT_FILE_PATHS.Where(filePath => !File.Exists(filePath)).Any() &&
+      !BACK_COMPAT_DIR_PATHS.Where(dirPath => !Directory.Exists(dirPath)).Any();
+  }
+
+  private static void ImportiOSNativeCompatibilityPackage() {
+    string iOSAudioLib = PLUGINS_IOS_PATH + IOS_AUDIO_LIB;
+    if (File.Exists(iOSAudioLib)) {
+      return;
+    }
+
+    string packagePath = Application.dataPath + IOS_NATIVE_COMPAT_PACKAGE_PATH;
+    AssetDatabase.ImportPackage(packagePath, true);
+    AssetDatabase.Refresh();
   }
 
   private static void ImportBackwardsCompatibilityPackage() {
@@ -139,16 +180,25 @@ public class GvrCompatibilityChecker {
   }
 
   private static void RemoveAnyBackwardsCompatibleFiles() {
-    IEnumerable<string> backCompatFiles = GetBackCompatFilePaths().AsEnumerable();
+    IEnumerable<string> backCompatFiles = BACK_COMPAT_FILE_PATHS.AsEnumerable();
     backCompatFiles = backCompatFiles.Where(filePath => File.Exists(filePath));
-    if (backCompatFiles.Count() == 0) {
+
+    IEnumerable<string> iOSBackCompatDirs = BACK_COMPAT_DIR_PATHS.AsEnumerable();
+    iOSBackCompatDirs = iOSBackCompatDirs.Where(dirPath => Directory.Exists(dirPath));
+
+    if (backCompatFiles.Count() == 0 && iOSBackCompatDirs.Count() == 0) {
       return;
     }
 
+    int dataPathLen = Application.dataPath.Length;
     string filesToRemove = "";
     foreach (string file in backCompatFiles) {
-      filesToRemove += string.Format("\t{0}\n", file);
+      filesToRemove += string.Format("\t{0}\n", file.Substring(dataPathLen + 1));
     }
+    foreach (string dir in iOSBackCompatDirs) {
+      filesToRemove += string.Format("\t{0}\n", dir.Substring(dataPathLen + 1));
+    }
+
     bool removeBackwardsCompatibleFiles = EditorUtility.DisplayDialog(
       BACK_COMPAT_FILES_FOUND_TITLE,
       string.Format("{0}{1}", BACK_COMPAT_FILES_FOUND_MESSAGE, filesToRemove),
@@ -159,11 +209,40 @@ public class GvrCompatibilityChecker {
 
     // Remove files.
     foreach (string file in backCompatFiles) {
-      AssetDatabase.DeleteAsset(ASSET_PATH_PREFIX + file.Substring(Application.dataPath.Length));
-      File.Delete(file);
-      File.Delete(file + META_EXT);
+      AssetDatabase.DeleteAsset(ASSET_PATH_PREFIX + file.Substring(dataPathLen));
+      if (File.Exists(file)) {
+        File.Delete(file);
+      }
+      if (File.Exists(file + META_EXT)) {
+        File.Delete(file + META_EXT);
+      }
       AssetDatabase.Refresh();
     }
+
+    // Remove iOS bundles.
+    foreach (string dir in iOSBackCompatDirs) {
+      AssetDatabase.DeleteAsset(ASSET_PATH_PREFIX + dir.Substring(dataPathLen));
+      // AssetDatabase may not fully delete files in versions < 5.6.
+      if (Directory.Exists(dir)) {
+        Directory.Delete(dir, true);
+      }
+      if (File.Exists(dir + META_EXT)) {
+        File.Delete(dir + META_EXT);
+      }
+      AssetDatabase.Refresh();
+    }
+  }
+
+  private static void RemoveiOSNativeIntegrationFiles() {
+    string iOSAudioLib = PLUGINS_IOS_PATH + IOS_AUDIO_LIB;
+    if (!File.Exists(iOSAudioLib)) {
+      return;
+    }
+    AssetDatabase.DeleteAsset(
+        ASSET_PATH_PREFIX + iOSAudioLib.Substring(Application.dataPath.Length));
+    File.Delete(iOSAudioLib);
+    File.Delete(iOSAudioLib + META_EXT);
+    AssetDatabase.Refresh();
   }
 
   private static bool IgnoreCompatibilityCheck() {
@@ -188,15 +267,6 @@ public class GvrCompatibilityChecker {
     }
 #endif  // UNITY_HAS_GOOGLEVR
     File.Create(IGNORE_MANIFEST_MERGE_CHECK_PATH);
-  }
-
-  private static string[] GetBackCompatFilePaths() {
-    return new string[] {
-      PLUGINS_ANDROID_PATH + ARMEABI_PATH + NATIVE_LIB,
-      PLUGINS_ANDROID_PATH + X86_PATH + NATIVE_LIB,
-      PLUGINS_ANDROID_PATH + COMMON_AAR,
-      PLUGINS_ANDROID_PATH + GVR_ACTIVITY_AAR,
-    };
   }
 }
 #pragma warning restore 414
